requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
SQLAlchemy>=2.0
psycopg2-binary>=2.9
pydantic>=2.6
pydantic-settings>=2.2
bcrypt>=4.1
PyJWT>=2.8
email-validator>=2.0